1st Semester/Python for Data Science/Unit4/PDS CT3/requirements.txt
plotly.expres
numerize